Rename `MutableReplacementList` to 'ReplacementList'
